renamed legend component
