+++ app/javascript/components/map/legend/VMapLegendEez.vue
